\name{breastrisk-package}
\alias{breastrisk-package}
\docType{package}
\title{
\packageTitle{breastrisk}
}
\description{
\packageDescription{breastrisk}
}
\details{
\packageDESCRIPTION{breastrisk}
\packageIndices{breastrisk}
An overview of how to use the package, including the most important functions
}
\author{
\packageAuthor{breastrisk}

Maintainer: \packageMaintainer{breastrisk}
}
\references{
Literature or other references for background information 
}
\keyword{ package }
%\seealso{
%~~ Optional links to other man pages, e.g. ~~
%~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
%}
\examples{
mybrm <- brm(50,25,3,4)

summary(mybrm)

}

